Add files via upload
pi_mpi - This is the executable
proj01.cpp - This is the source code
proj01.sh - This is the shell code that activates and sends the executable to the different nodes
PI_MPI_XXX.log - These are the log files for different variables. How many nodes, and how many random values to create.
proj01_log (1).xlsx - Data table containing the summarized results of all of the logs.